Don't use the schema name when querying repmgr for failover databases
The schema name is dependent on the repmgr cluster name which we are
planning on basing on the region number of the primary database.

When registering the primary node we will add the new schema to the
root user's search_path so that the repmgr tables can be queried
without specifying the schema explicitly.

(transferred from ManageIQ/manageiq-gems-pending@8fb17bffdfdb07b06fed94018eef2d1efe722e86)