Merge connection info from failover.yml with existing database.yml info
This will allow things like the password, which will not be stored
in the repmgr schema, to be used when connecting to the database to
check that it is not in recovery mode.

Without this change a typical setup (with password authentication
between the database server and the rails app) will fail when
attempting to failover because we will not have the password to connect
to the standby databases.

(transferred from ManageIQ/manageiq-gems-pending@f8289baedace097f76e2396527425a862b4c59c9)